added OpenMW patch
Sitter Brelyna is now spawned by a leveled list in the core mod and the spawn rate is 100.
In the OpenMW Patch leveled list spawn rate is 0 but a normal standing Brelyna is next to the stair.
This method has a downside where she will respawn but I don't want to make a plugin replacer for such a small issue.